Gem Bug Fix
Changed:
Changed the collision setting for the gem flipbook to "All Frames
Collision" instead of no collision.
Changed the overlap function of gem from overlapping a sprite to
overlapping a flipbook. (The gem object is a flipbook and no longer a
single sprite).

Other Fixes:
Fixed compiler error in wave 5.

Contributors: Michael/Stuart